2007-06-18  Jeffrey Stedfast  <fejj@novell.com>
	* brush.cpp (VideoBrush::SetupBrush): Get a reference to the
	MediaElement if we don't already have one.
	(VideoBrush::OnPropertyChanged): SourceName does not refer to a
	uri, instead SourceName refers to a MediaElement.

	* media.cpp: Implemented some more helper "setters", defaulted
	some more MediaElement property values to correct values, and
	implemented updating of both the CurrentState property and the
	PositionProperty.

	* mplayer.cpp (Position): Implemented.


svn path=/trunk/moon/; revision=80051

+++ src/mplayer.h
@@ -62,8 +62,8 @@ public:
 	void Stop ();
 	
 	bool CanSeek ();
-	void Seek (double position);
-	double Position ();
+	void Seek (int64_t position);
+	int64_t Position ();
 	
 	void Mute ();
 	void UnMute ();
